requirements.txt
requests==2.31.0
python-dotenv==1.0.0
boto3==1.26.138
pylint==2.17.4
black==23.3.0
isort==5.12.0
bandit==1.7.5
-r ./app/requirements.txt